requirements.txt
streamlit
Pillow
numpy
faiss-cpu
open_clip_torch
tqdm
